refactor: `isProduction` 유틸함수화
