+++ packages/web/src/modules/i18n/type.ts
@@ -1,11 +1,9 @@
-import { languages } from '.';
+import { languages } from './constant';
 
 export type Language = (typeof languages)[number];
-
 export type LanguageType = {
   lng: Language;
 };
-
 export interface LanguageParams {
   params: LanguageType;
 }
