+++ packages/web/src/utils/Environment.ts
@@ -1,0 +1,1 @@
+export const isProduction = process.env.NODE_ENV === 'production';
