+++ packages/web/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './cn';
+export * from './Environment';
